# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff/sync-jira:
    post:
      operationId: syncStaffJira
      tags: [operations]
      summary: Sync staff shifts from Jira
      description: Runs the server-side Jira shift synchronizer and returns the number of processed shifts.
      responses:
        "200":
          description: Jira shift synchronization result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JiraSyncResult"
        "503":
          description: Jira synchronization is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /staff/{id}:
    patch:
      operationId: updateStaffStatus
      tags: [operations]
      summary: Update a staff member's status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffStatusUpdate"
      responses:
        "200":
          description: Updated staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
...
  /release-gates/{id}/check:
    patch:
      operationId: toggleReleaseGate
      tags: [operations]
      summary: Check or uncheck a release gate item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated release gate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseGate"
  /procurement:
    get:
      operationId: listProcurementRecords
      tags: [operations]
      summary: List procurement workflow records
      responses:
        "200":
          description: Procurement workflow records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProcurementRecord"
    post:
      operationId: createProcurementRecord
      tags: [operations]
      summary: Create a purchase requisition (PR)
      description: Validates budget availability and auto-sets legal/security review flags (HKD > 100,000).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcurementInput"
      responses:
        "201":
          description: Created procurement record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcurementRecord"
        "409":
          description: Budget pre-check failed or duplicate PR number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /procurement/{id}/approve:
    patch:
      operationId: approveProcurement
      tags: [operations]
      summary: Approve a procurement record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated procurement record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcurementRecord"
  /procurement/{id}/status:
    patch:
      operationId: advanceProcurementStatus
      tags: [operations]
      summary: Advance PR/PO through tiered status transitions
      description: Enforces tiered approval matrix (L1/L2/L3), legal/security review gating, and dual-control.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvanceProcurementInput"
      responses:
        "200":
          description: Updated procurement record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcurementRecord"
        "409":
          description: Transition rejected (review gating / approval tier / variance)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /procurement/{id}/review:
    patch:
      operationId: submitProcurementReview
      tags: [operations]
      summary: Submit legal or security review decision
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecisionInput"
      responses:
        "200":
          description: Updated procurement record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcurementRecord"
        "409":
          description: Review not required for this record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /procurement/{id}/payments:
    get:
      operationId: listPaymentSchedules
      tags: [operations]
      summary: List payment schedules for a procurement record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentSchedule"
    post:
      operationId: createPaymentSchedule
      tags: [operations]
      summary: Create a milestone / payment schedule entry
      description: Standard 3:4:3 milestone structure; auto-triggers three-way match on invoice upload.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentInput"
      responses:
        "201":
          description: Created payment schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSchedule"
        "409":
          description: Milestones would exceed PO amount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/{id}/invoice:
    patch:
      operationId: submitInvoice
      tags: [operations]
      summary: Submit vendor invoice or OCR data, triggering three-way match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "200":
          description: Payment schedule with three-way match result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSchedule"
  /payments/{id}/three-way:
    get:
      operationId: getThreeWayMatch
      tags: [operations]
      summary: Get the three-way match result for a payment schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Three-way match result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreeWayMatch"
        "404":
          description: No three-way match found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/{id}/variance:
    patch:
      operationId: resolveVariance
      tags: [operations]
      summary: Resolve a blocked payment variance (finance + legal consultation)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VarianceResolutionInput"
      responses:
        "200":
          description: Resolved payment schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSchedule"
        "409":
          description: No variance to resolve or not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/{id}/signoff:
    patch:
      operationId: dualSignoff
      tags: [operations]
      summary: Apply dual sign-off (Head of IT + Finance Auditor) for payments > 250k
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignoffInput"
      responses:
        "200":
          description: Signed-off payment schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSchedule"
        "409":
          description: Dual sign-off not required or incomplete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    StaffMember:
      type: object
      required: [id, name, initials, role, team, region, status, ticket, environment, eta, updatedAt, isStale]
      properties:
        id: { type: string }
        name: { type: string }
        initials: { type: string }
        role: { type: string }
        team: { type: string }
        region: { type: string }
        status: { type: string }
        ticket: { type: string }
        environment: { type: string }
        eta: { type: string }
        updatedAt: { type: string }
        isStale: { type: boolean }
    StaffStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }
    ReleaseGate:
      type: object
      required: [id, environment, title, owner, due, checked, risk]
      properties:
        id: { type: string }
        environment: { type: string }
        title: { type: string }
        owner: { type: string }
        due: { type: string }
        checked: { type: boolean }
        risk: { type: string }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        category:
          type: string
          enum: [CONFIGURATION, JIRA, SUPABASE, UNKNOWN]
    JiraSyncResult:
      type: object
      required: [count]
      properties:
        count: { type: number, minimum: 0 }
...
    AdvanceProcurementInput:
      type: object
      required: [toStatus, actorId]
      properties:
        toStatus:
          type: string
          enum: [PR_APPROVED, PO_ISSUED, MILESTONE_RECEIVED, INVOICE_PENDING, PAYMENT_APPROVED, PAID]
        actorId: { type: string }
        note: { type: string }
...